Moving fixtures into seprate file
